[FIX] Use default destructor.
Making the destructor wait until the queue is empty is an ill-formed
behavior. The non-empty queue should be destructable if no reader or
writer are modifying the queue. This needs to be guarded with external
synchonisation points.
Also lifts the requirement on types to be convertible to the queues
value type and not to be the same anymore.