[FIX] Preserve the inherited operator of the base iterator.
This changes fixes general expectations about the inherited iterator base.
Before this change it was implementing almost all operators by means of += and ++.
However, the base should take care of the correct implementation of each operator and
the behavior of these operations should be preserved.
In addition, this commit removes the concept requirements of each operation and
solely checks if the current operator is supported by the base class.
Accoringly, the derived iterator inherits the actual functionality which
must not be related to any iterator concept, e.g. when inheriting operators from a
pseudo random access iterator.