[MISC] Pass callback into the alignment algorithm and call it on every alignment result.
This allows us to define the alignment result type only once at the top-level when the algorithm is configured.
Accoridingly, the type must not be redefined within the algorithms and possible incompatibilities between the calls are avoided.
In addition, the algorithm can now decide when to call the user callback which will allow in the future to signal back to the algorithm what should happen.
For example interrupt the execution.