[MISC] use std::(ranges::range|iter)_(value|reference)_t
This fixes part of #1549. I'm not sure why this version works, but I
tested it for gcc 7, 8 and 9 and worked.

The main problem I had was to allow the syntax:

```c++
// this should call the constructor which takes two iterators
concatenated_sequences seqs{iterator, sentinel};
```

https://github.com/seqan/seqan3/blob/75f934f1886ee225fb1bad01b9c9f474caf4c508/include/seqan3/range/container/concatenated_sequences.hpp#L299-L307

In many versions I tried, this call instead tried

https://github.com/seqan/seqan3/blob/75f934f1886ee225fb1bad01b9c9f474caf4c508/include/seqan3/range/container/concatenated_sequences.hpp#L321-L328

This caused some issues, because the template also tried to instantiate
those constructors with not-yet defined class types (like for example the
iterator type within the class)

All those complicated overloads are only available in this class, to
allow constructions of sequences of "compatible" type (i.e. range value
types that are convertible to the value type of the class)
`std::vector<rna4>` and `std::vector<dna4>`.

For the future (and maybe gcc10 compatibility) we should remove the
ability to plug any possible way to add somehow compatible types. This
should be limited.

For now the proposed fix keeps the old behaviour.