[TEST] PrintTo overload is bad
If you include `seqan3/test/pretty_printing.hpp` it can make a valid existing
test case not work any more. The reason is that `pretty_printing`, which
uses the `debug_stream` internally, will overrule the default implementation
of google. This is a problem if google supports a type (like in this example
an iterator), but the debug_stream doesn't. But the debug_stream says it
supports any type, so we might have cases where a test could be written
before, but will fail after adding `pretty_printing`.