[FIX] gcc10: use std::variant::emplace instead of implicit assignment
Our PPC CI uses 10.1.0 release from source and it triggered an error
which most distros seems to have hot-patched.

I encountered that problem in the past and used this workaround, but the
problem was fixed in the meantime so that I did not need to submit this
patch.

The patch is confirmed to work by @eseiler