[FIX] a better fix for concatenated sequences?!
This could have been the fix for #1731, but I just recently saw this
technique in the standard library. This solution should be a lot less
janky than the previous one. I tested it with gcc7, 8, 9 and 10 and it
worked without a problem.

I wanted to finish #1549 and after changing the definition of
`dimension_v` the previous construction of #1731 broke, so had to
rethink a solution for this.