[FEATURE] Adds refurbished policy for tracking the optimum of the alignment.
This new policy is not a CRTP base class anymore but a regular base class with protected member functions such that only the derived class can instantiate this policy.
The behavior of the tracking functions changed.
Now every cell is passed through and possibly read to update the optimum if the configuration requires it.