[FIX] cxx20: use std::default_sentinel_t on minimiser view
minimiser_hash_test did not compile with `gcc10 -std=c++2a`

---

The expression `std::ranges::begin(this->test_range) == std::ranges::cend(this->test_range)`
for `this->test_range = seqan3::detail::minimiser_view` did not compile.

I tried to find out, but ultimately could only figure out that
the left hand side was

* `lhs_iterator = std::ranges::begin(this->test_range)`
* `lhs_iterator == minimiser_view<rng>::window_iterator<rng>`
  * with `rng = non-const std::ranges::transform_view` internally defined as
  * `lhs_iterator::sentinel == std::ranges::end(rng)` and
  * `std::ranges::end(rng) == std::ranges::transform_view::sentinel<false>`

Whereas the right side was

* `rhs_sentinel = std::ranges::cend(this->test_range)`
* `rhs_sentinel == std::ranges::cend(rng)`
    * with `rng = non-const std::ranges::transform_view`
* `rhs_sentinel == std::ranges::transform_view::sentinel<true>`

The `==` in `lhs_iterator` is defined as
* `operator==(lhs_iterator const &, lhs_iterator::sentinel const &)` which is the same as
* `operator==(minimiser_view<rng>::window_iterator<rng> const &, std::ranges::transform_view::sentinel<false> const &)`

The requested `==` of the expression
* `std::ranges::begin(this->test_range) == std::ranges::cend(this->test_range)`
* needs the following signature
* `operator==(minimiser_view<rng>::window_iterator<rng> const &, std::ranges::transform_view::sentinel<true> const &)`

As you can see, the signature does not match completely.

Normally you can convert a `std::ranges::transform_view::sentinel<false>` into a
`std::ranges::transform_view::sentinel<true>` and from what I can see, the compiler
is normally able to do this conversion sequence.

See [std::ranges::transform_view::iterator](https://eel.is/c++draft/range.transform.iterator)
and [std::ranges::transform_view::sentinel](https://eel.is/c++draft/range.transform.sentinel).

They both don't have a special case for `sentinel<false>` or `sentinel<true>`.
(Note that they have less comparison operators, this is because since std20 you automatically define all reverse
operators for `==` e.g. `a == b` also defines `b == a` and you get `!=` for free if `==` was defined.)

minimiser view has the same comparison operators defined and has a constructor for non-const to const conversion, but
fails to find the conversion sequence.

My guess is that the conversion sequence does not work, because lhs and rhs are in different namespaces.
(Haven't confirmed that yet)

---

The fix in this PR is to make the signature of the `==` operator the same for const and non-const version.

This is done by using `std::default_sentinel_t`.

It turned out that this is a defect in the standard, see
* https://gcc.gnu.org/bugzilla/show_bug.cgi?id=95322
* https://cplusplus.github.io/LWG/issue3448