[FEATURE] Adjust the search range and search configuration to use algorithm executor.
Updates the two search algorithms be invoked with a second callback argument.
This callback is invoked with the generated search results.
Also the search algorithm now associates an index
with every query and propagates this indexed query pair to the algorithms.
Updates the search result range to work with the algorithm executor.