[FIX] Explicit return type in transform lambda to avoid hard compiler error.
When the return type for the lambda is not given the entire body must be evaluated in the overload resolution of the transform view.
There also the const qualified version is checked by the compiler. If inside the body of the lambda function is
an expression that results in a constraint error the compiler will omit an hard error because he cannot deduce the return type.
To fix this, the return type is given explicitly in order to allow SFINAE friendly substituion errors due to incomplete types in the return type.