[FEATURE] Adds a combined alignment matrix.
This matrix combines the score matrix with the trace matrix and provides a unified alignment column for the alignment algorithm.
Thus, the algorithm can be implemented based on the affine_cell_proxy type and offer additional
interfaces depending on the properties, i.e. switching between computing the trace or not.