[MISC] Prepare undbanded edit distance for new align_cfg::output
With the new align_cfg::output configuration the different features that are computed are not dependent anymore.
However, some computations of the edit distance algorithm are necessary in order to compute a specific feature. For example, when the alignment is requested the edit distance algorithm must also compute the begin and end positions.
This commit separates the edit distance configuration from the user configurations,
i.e. if the user requested an alignment but not the begin/end positions,
the algorithm still executes correctly, while at the end only the alignemnt will be stored.