[FEATURE] Add correct padding symbol for simd scoring scheme.
Depending on the selected simd scoring scheme different padding symbols need to be used.
Defining this in the traits class was the wrong place, since we cannot determine the simd score type from the alignment configuration.
Instead, the simd scoring scheme itself should expose the correct padding symbol since it is implementation dependent.
It also fixes the padding symbol for global and local as well as the padded match score.
The current implementation does not work with 8 bit score yet. This is another feature implemented later.