[TEST] add direct align_pairwise versions for edit_distance
Normally, we use align_pairwise() to auto-detect which alignment
algorithm should be used. In the edit distance benchmarks we have a
benchmark that compares the cost of directly calling the edit distance
algorithm to the align_pairwise call.

For this I wrote (for seqan2 and seqan3) a helper function to call the
algorithm directly.