[MISC] Simplify the pipe operator for the configurations.
This merges the pipe-operator overloads into a single free function, which is implemented by means of the new append function.
Futher, the pipe-operator uses concepts for to check for eligible types and does not cast to the base class of the configuration elements anymore.

The pipeability test in the pipeable_config_test_template had to be removed since it was testing an invalid scenario.
Later a better test template infrastructure will be introduced.