[MISC] introduce [rd]na4::char_to_rank function
I tried to define the lookup tables within the function itself, but we
can't declare the lookup table as `static constexpr ...`, because the
function is declared constexpr and does not allow any linkage within it.

So we would need to declare the lookup table as constexpr only, like so

```patch
-    static constexpr char_type rank_to_char[alphabet_size]
+    static constexpr char_type rank_to_char(rank_type const rank)
+    {
+        constexpr char_type lookup_table[] =
         {
             'A',
             'C',
             'G',
             'T'
         };

+        return lookup_table[rank];
+    }
```

The problem is that gcc does not optimise (clang does!) this case with
lookup tables, but moves the memory-in every time we call this function.

See https://godbolt.org/z/K8fcqe.

The micro benchmark `test/performance/alphabet/alphabet_assign_char_benchmark`
showed that `assign_char<seqan3::dna4>` took `81.6 ns` before that
patch, but `2903 ns` after that change.

So I reverted to use the normal in-class lookup-table declaration that is now
used within the static function itself.

I would have preferred to remove that class-visible definition.

See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=99320