[FEATURE] add class test::tmp_directory
This class is a convenience class that creates a temporary directory.
This temporary directory is secured by sandboxed_path lowering the
likelihood of leaving the temporary directory by accident.
Additionally it provides functionality like automatic cleanup of the temporary
directory, but also functionality to check if the directory is empty.