[FIX] correct parsing of IDs in FASTA files
FASTA has the following structure:
```
> MyID
MySeq
```

This should be parsed into 'MyID' and 'MySeq' (This works fine).

If the sequence id has '>' as a character, this might fail:

```
> >MyID
MySeq
```
This will be parsed into 'MyId' and 'MySeq', which does not meet the
specification. It should be parsed to '>MyID' and 'MySeq'.

This PR fixes this.