[FIX] Alignment Wrong alignment score #3043
Added carry bit to up and left open trace directions in order to capture original open signal even if it is overwritten by the orhtogonal direction.
In the particular issue the left extension overrites the up open signal in the last column and fifth row.
The traceback now follows the vertical or horizontal direction unitl it encounters the carry bit signal.