[FEATURE] Switch to the new hamming distance scoring scheme to configure Myers' bitparallel edit distance computation.
Now the edit distance is selected internally if the hamming scheme was used explicitly and if the gap scheme gives -1 for extending and 0 for opening a gap.
The edit distance can now be used independent of the alphabet type and is not bound to seqans nucleotide scoring schemes.
Since the recommended way to enable the fast bitparallel algorithm is to use the `align_cfg::edit_scheme` configuration element, the majority of user code should work without noticing the internal changes made to the configuration.