[FIX] Fixes wrong default for internal affine gap policy in alignment algorithm.
During configruation we check if the gap scheme is set and if not use the default gap scheme with `gap open = 0` and `gap extension = -1`. But this default gap scheme is ignored and was later set to -10 and -1 in the alignment gap policy.
This commit fixes this, to keep the defaults consistent.