[INFRA] Make versions cache variables
Exemplary use case: Installed benchmark version is 1.8.3.
CPM will not use 1.8.3, because 1.8.3 < 1.9.1.

Making the versions CACHE allows setting them on the command line.
Making them FORCE means that the value given on the command line is not
cached. The versions has to be provided for each CMake call, and will
revert to the default value otherwise.