[FIX] gcc15: conditionally borrowed_range views
gcc15 now implements std::ranges::to
For `std::string` (as example), the call is delegated to
`std::string(std::from_range, R &&)`.
This then will use `std::ranges::data` to get the data pointer and just
copy the data.
This happens, for example, in format_sam's
`id | detail::take_until_and_consume(is_space) | ranges::to<id_type>()`.
Since the view is not an lvalue, it needs to be a borrowed_range for
`std::ranges::data` to work.
Similar to std::views, we make (some of) our views conditionally
borrowed if the underlying range is borrowed.